Convert parse environemnt to parse authorization
